refactor(web/prototype-fixtures): split lib/prototype-fixtures.ts into a directory
The monolith (515 LOC) violated the 350-LOC project rule. Splits
along the natural seams already marked by the section banners in
the original file:

  types.ts           — every export interface (167 LOC)
  data.ts            — raw JSON load + publicVisible + hotScore (39)
  submissions.ts     — submission/comment/user queries (123)
  tags.ts            — tag queries (34)
  projects.ts        — project queries (33)
  notifications.ts   — mock notification feed (57)
  mod-queue.ts       — mock mod-queue + audit fixtures (121)
  index.ts           — barrel of the public surface (60)

Eleven existing importers of `@/lib/prototype-fixtures` resolve
through the barrel unchanged. No behavioral change.

Mock mod-queue + audit fixtures get a comment in mod-queue.ts noting
they will be replaced by reads against decision_records (editorial)
and policy_decisions (policy moderator) once those production read
paths exist. They are not connected to either today.
